pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "pysdmx-lbs-conversion"
version = "0.1.0"
description = "Legacy BIS LBS line format to SDMX 3.x generator using pysdmx"
authors = [{name="Your Name", email="you@example.org"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "pysdmx[xml]>=0.1.0",  # Adjust to the actual version of your fork/published package
  "pyyaml>=6.0"
]

[project.optional-dependencies]
pandas = ["pandas>=2.0"]

[tool.black]
line-length = 100